Improve initGitRepo function to create .gitignore only if it doesn't exist

+++ shared/zsh-config.sh
@@ -154,11 +154,13 @@ function generatePersonalSshKey() {
 }
 
 function initGitRepo() {
-    # Create .gitignore
-    if [ "$2" = "python" ]; then
-        echo ".venv" > .gitignore
-    else
-        touch .gitignore
+    # Create .gitignore if it doesn't exist
+    if [ ! -f .gitignore ]; then
+        if [ "$2" = "python" ]; then
+            echo ".venv" > .gitignore
+        else
+            touch .gitignore
+        fi
     fi
 
     git init
